requirements.txt
beautifulsoup4==4.11.1
loguru==0.6.0
openai==0.26.0
pipreqs==0.4.11
pyperclip==1.8.2
python-dotenv==0.21.0
requests==2.27.1
setuptools==65.6.3
snakeviz==2.1.1
winotify==1.1.0
